Organize the module into multiple files.
